- pu write frames to disk
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@2460 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/plugins/services/mediaexporter/FileExporter.cpp
@@ -204,7 +204,8 @@ namespace org {
            * clean up temporary files, they are no longer needed
            */
           fis.close();
-          infile.deleteFile();
+          if(!_source_stream_map[pu._output_packets.front()->getStreamIndex()].stream_type == AVMEDIA_TYPE_VIDEO)
+                infile.deleteFile();
         }
         if (true) {
           db::JobLog log(job.getDatabase());
